requirements.txt
streamlit==1.35.0
langchain==0.2.0
langchain_ollama==0.0.11
langchain_chroma==0.1.0
pymupdf==1.24.2
pytesseract==0.3.10
requests==2.31.0
python-dotenv==1.0.1
nomic==4.0.0    
Pillow==10.3.0